# specifications/openapi.yaml
...
openapi: 3.0.2
info:
  description: >
    A self guided random access training system on subjects related to identifying
    and evaluating opportunities to improve industrial energy, 
    resource and production efficiency.

  version: 1.0.0
  title: EEC Walkthrough API
  contact:
    name: API support
    email: 'thomasza@oregonstate.edu'
...
paths:
...
  /Headers:
    post:
      summary: Creates a new header
      description: >
        Creates a new header
      operationId: createHeader
      tags:
        - Headers
      requestBody:
        $ref: '#/components/requestBodies/Header_POST'
      responses:
        '201':
          $ref: '#/components/responses/Header_POST_201'
        '403':
          $ref: '#/components/responses/403'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  '/Headers/{headerId}':
    get:
      summary: Gets a header by ID
      description: >
        Gets a header using the header ID
      operationId: getHeader
      tags:
        - Headers
      parameters:
        - $ref: '#/components/parameters/Header_headerId'
      responses:
        '200':
          $ref: '#/components/responses/Header_GET_200'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

    patch:
      summary: Update a header by ID
      description: >
        Updates one or more header fields
      operationId: updateHeader
      tags:
        - Headers
      requestBody:
        $ref: '#/components/requestBodies/Header_PATCH'
      parameters:
        - $ref: '#/components/parameters/Header_headerId'
      responses:
        '200':
          $ref: '#/components/responses/PATCH_200'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'

    delete:
      summary: Delete a header by ID
      description: >
        Delete a header using the header ID
      operationId: deleteHeader
      tags:
        - Headers
      parameters:
        - $ref: '#/components/parameters/Header_headerId'
      responses:
        '200':
          $ref: '#/components/responses/DELETE_200'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
  /Cards:
    post:
      summary: Creates a new card
      description: >
        Creates a new card
      operationId: createCard
      tags:
        - Cards
      requestBody:
        $ref: '#/components/requestBodies/Card_POST'
      responses:
        '201':
          $ref: '#/components/responses/Card_POST_201'
        '403':
          $ref: '#/components/responses/403'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  '/Cards/{cardId}':
    get:
      summary: Gets a card by ID
      description: >
        Gets a card using the card ID
      operationId: getCard
      tags:
        - Cards
      parameters:
        - $ref: '#/components/parameters/Card_cardId'
      responses:
        '200':
          $ref: '#/components/responses/Card_GET_200'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

    patch:
      summary: Update a card by ID
      description: >
        Updates one or more card fields
      operationId: updateCard
      tags:
        - Cards
      requestBody:
        $ref: '#/components/requestBodies/Card_PATCH'
      parameters:
        - $ref: '#/components/parameters/Card_cardId'
      responses:
        '200':
          $ref: '#/components/responses/PATCH_200'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'

    delete:
      summary: Delete a card by ID
      description: >
        Delete a card using the card ID
      operationId: deleteCard
      tags:
        - Cards
      parameters:
        - $ref: '#/components/parameters/Card_cardId'
      responses:
        '200':
          $ref: '#/components/responses/DELETE_200'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
  /Items:
    post:
      summary: Creates a new item
      description: >
        Creates a new item
      operationId: createItem
      tags:
        - Items
      requestBody:
        $ref: '#/components/requestBodies/Item_POST'
      responses:
        '201':
          $ref: '#/components/responses/Item_POST_201'
        '403':
          $ref: '#/components/responses/403'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'
  '/Items/{itemId}':
    get:
      summary: Gets an item by ID
      description: >
        Gets an item using the item ID
      operationId: getItem
      tags:
        - Items
      parameters:
        - $ref: '#/components/parameters/Item_itemId'
      responses:
        '200':
          $ref: '#/components/responses/Item_GET_200'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

    patch:
      summary: Update an item by ID
      description: >
        Updates one or more item fields
      operationId: updateItem
      tags:
        - Items
      requestBody:
        $ref: '#/components/requestBodies/Item_PATCH'
      parameters:
        - $ref: '#/components/parameters/Item_itemId'
      responses:
        '200':
          $ref: '#/components/responses/PATCH_200'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
        '500':
          $ref: '#/components/responses/500'

    delete:
      summary: Delete an item by ID
      description: >
        Delete an item using the item ID
      operationId: deleteItem
      tags:
        - Items
      parameters:
        - $ref: '#/components/parameters/Item_itemId'
      responses:
        '200':
          $ref: '#/components/responses/DELETE_200'
        '401':
          $ref: '#/components/responses/401'
        '500':
          $ref: '#/components/responses/500'
components:
  schemas:
...
    Header:
      description: An object representing a specific page header
      type: object
      required:
        - headerId
        - pageId
        - title
        - userId
        - created
        - approved
      properties:
        headerId:
          allOf:
            - description: The unique identifier of the header
            - $ref: '#/components/schemas/Id'
        pageId:
          allOf:
            - description: The unique identifier of the page that this header belongs to
            - $ref: '#/components/schemas/Id'
        title:
          description: The title of the header
          type: string
          minLength: 1
          maxLength: 100
          example: Compressed Air Opportunities to Consider
        userId:
          allOf:
            - description: The user ID of the user who created the header
            - $ref: '#/components/schemas/Id'
        created:
          allOf:
            - description: The time that the header was created
            - $ref: '#/components/schemas/UtcTime'
        approved:
          description: Value indicating if the header has been approved to be shown on the website
          type: boolean
          minimum: 0
          maximum: 1
          example: 0
    Card:
      description: An object representing a specific card that belongs to a header
      type: object
      required:
        - cardId
        - headerId
        - title
        - userId
        - created
        - approved
      properties:
        cardId:
          allOf:
            - description: The unique identifier of the card
            - $ref: '#/components/schemas/Id'
        headerId:
          allOf:
            - description: The unique identifier of the header that the card belongs to
            - $ref: '#/components/schemas/Id'
        title:
          description: The title of the card
          type: string
          minLength: 1
          maxLength: 100
          example: Best Practices
        userId:
          allOf:
            - description: The user ID of the user who created the card
            - $ref: '#/components/schemas/Id'
        created:
          allOf:
            - description: The time that the card was created
            - $ref: '#/components/schemas/UtcTime'
        approved:
          description: Value indicating if the card has been approved to be shown on the website
          type: boolean
          minimum: 0
          maximum: 1
          example: 0
    Item:
      description: An object representing a specific item that belongs to a card
      type: object
      required:
        - itemId
        - cardId
        - parentId
        - iconType
        - contentText
        - contentUrl
        - contentLabel
        - userId
        - created
        - approved
      properties:
        itemId:
          allOf:
            - description: The unique identifier of the item
            - $ref: '#/components/schemas/Id'
        cardId:
          allOf:
            - description: The unique identifier of the card that the item belongs to
            - $ref: '#/components/schemas/Id'
        parentId:
          allOf:
            - description: The ID of the parent item that this item belongs to. The child item of a parent is indented relative to the parent's level of indentation. Defaults to NULL if there is no parent item.
              type: integer
              format: int32
              minimum: 1
              maximum: 4294967295
              example: 15
        iconType:
          description: The integer value that reprsents a specific item type. The value can then be referenced from a table on the frontend to determine what icon to show and how to display the items content.
          type: integer
          format: int16
          minimum: 1
          maximum: 65535
          example: 5
        contentText:
          description: The main body of text of the item.
          type: string
          minLength: 0
          maxLength: 1000
          example: This guide focuses mainly on screw and reciprocating compressors. These are the most common...
        contentUrl:
          description: If a URL is relevant to this item then this URL points to an image or resource, otherwise this is an empty string
          type: string
          minLength: 0
          maxLength: 1000
          example: https://www.compressedairchallenge.org/data/sites/1/media/library/sourcebook/Improving_Compressed_Air-Sourcebook.pdf
        contentLabel:
          description: If this item is linking to a resource, then this is the string used to represent the link. It can also be a title, otherwise this is an empty string
          type: string
          minLength: 0
          maxLength: 1000
          example: Assessing Industrial Air Compressor
        userId:
          allOf:
            - description: The user ID of the user who created the item
            - $ref: '#/components/schemas/Id'
        created:
          allOf:
            - description: The time that the item was created
            - $ref: '#/components/schemas/UtcTime'
        approved:
          description: Value indicating if the item has been approved to be shown on the website
          type: boolean
          minimum: 0
          maximum: 1
          example: 0
    Id:
      type: integer
      format: int32
      minimum: 1
      maximum: 4294967295
      example: 15
...
    UtcTime:
      type: string
      format: 'yyyy-mm-dd HH:MM:SS'
      minLength: 19
      maxLength: 19
      example: '2020-12-31 23:59:59'
...
  requestBodies:
...
    Header_POST:
      required: true
      content:
        application/json:
          schema:
            properties:
              pageId:
                allOf:
                  - $ref: '#/components/schemas/Id'
              title:
                example: Compressed Air Opportunities to Consider
              userId:
                allOf:
                  - $ref: '#/components/schemas/Id'
    Header_PATCH:
      required: true
      content:
        application/json:
          schema:
            properties:
              pageId:
                allOf:
                  - description: The unique identifier of the page that this header belongs to
                  - $ref: '#/components/schemas/Id'
              title:
                description: The title of the header
                type: string
                minLength: 1
                maxLength: 100
                example: Compressed Air Opportunities to Consider
              approved:
                example: 1
    Card_POST:
      required: true
      content:
        application/json:
          schema:
            properties:
              headerId:
                example: 25
              title:
                example: Rules of Thumb
              userId:
                example: 10
    Card_PATCH:
      required: true
      content:
        application/json:
          schema:
            properties:
              headerId:
                example: 25
              title:
                example: Rules of Thumb
              approved:
                example: 1
    Item_POST:
      required: true
      content:
        application/json:
          schema:
            properties:
              cardId:
                example: 35
              parentId:
                example: 40
              iconType:
                example: 5
              contentText:
                example: This guide focuses mainly on screw and reciprocating compressors. These are the most common...
              contentUrl:
                example: https://www.compressedairchallenge.org/data/sites/1/media/library/sourcebook/Improving_Compressed_Air-Sourcebook.pdf
              contentLabel:
                example: Assessing Industrial Air Compressor
              userId:
                example: 15
    Item_PATCH:
      required: true
      content:
        application/json:
          schema:
            properties:
              cardId:
                example: 35
              parentId:
                example: 40
              iconType:
                example: 5
              contentText:
                example: This guide focuses mainly on screw and reciprocating compressors. These are the most common...
              contentUrl:
                example: https://www.compressedairchallenge.org/data/sites/1/media/library/sourcebook/Improving_Compressed_Air-Sourcebook.pdf
              contentLabel:
                example: Assessing Industrial Air Compressor
              approved:
                example: 1
  parameters:
...
    Header_headerId:
      name: headerId
      in: path
      description: The unique identifier of the header
      required: true
      schema:
        $ref: '#/components/schemas/Header/properties/headerId'
    Card_cardId:
      name: cardId
      in: path
      description: The unique identifier of the card
      required: true
      schema:
        $ref: '#/components/schemas/Card/properties/cardId'
    Item_itemId:
      name: itemId
      in: path
      description: The unique identifier of the item
      required: true
      schema:
        $ref: '#/components/schemas/Item/properties/itemId'
  responses:
...
    PATCH_200:
      description: An object containing information about the changed rows
      content:
        application/json:
          schema:
            properties:
              changedRows:
                example: 3
...
    DELETE_200:
      description: The number of rows deleted
      content:
        application/json:
          schema:
            description: An object containing the number of rows deleted
            type: object
            required:
              - Id
            properties:
              affectedRows:
                type: integer
                example: 1
    Header_POST_201:
      description: Header created
      content:
        application/json:
          schema:
            description: An object containing information about the new header
            type: object
            required:
              - Id
            properties:
              insertId:
                $ref: '#/components/schemas/Id'
    Header_GET_200:
      description: An object containing information about the header
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Header'
    Card_POST_201:
      description: Card created
      content:
        application/json:
          schema:
            description: An object containing information about the new card
            type: object
            required:
              - Id
            properties:
              insertId:
                $ref: '#/components/schemas/Id'
    Card_GET_200:
      description: An object containing information about the card
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Card'
    Item_POST_201:
      description: Item created
      content:
        application/json:
          schema:
            description: An object containing information about the new item
            type: object
            required:
              - Id
            properties:
              insertId:
                $ref: '#/components/schemas/Id'
    Item_GET_200:
      description: An object containing information about the item
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Item'
...
    '401':
      description: Missing or invalid authentication credentials
      content:
        application/json:
          schema:
            properties:
              error:
                description: A message describing the error
                type: string
                minLength: 1
                example: Unauthorized.
    '403':
      description: Not allowed to perform this action
      content:
        application/json:
          schema:
            properties:
              error:
                description: A message describing the error
                type: string
                minLength: 1
                example: Forbidden.
    '404':
      description: Specified resource not found
      content:
        application/json:
          schema:
            properties:
              error:
                description: A message describing the error
                type: string
                minLength: 1
                example: Not found.
    '422':
      description: Unprocessable entity
      content:
        application/json:
          schema:
            properties:
              errors:
                description: A list of input validation errors
                type: array
                example: [
                      {
                          "value": "",
                          "msg": "Invalid value",
                          "param": "userName",
                          "location": "body"
                      },
                      {
                          "value": "",
                          "msg": "Invalid value",
                          "param": "email",
                          "location": "body"
                      }
                  ]
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            properties:
              error:
                description: A message describing the error
                type: string
                minLength: 1
                example: An internal server error occurred. Please try again later.

# Added by API Auto Mocking Plugin
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/silverware13/EEC_Walkthrough/1.0.0